Move workspaces list to separate file
